Split kraken module into kraken / krona / kraken_builder
